requirements.txt
telebot==0.0.4
Pillow==8.0.1
requests==2.25.1